src: move custom messages to messages.js
Unified-list don't needs to be responsible
to print custom messages.

Connects to #211
