+++ lib/messages.js
@@ -1,9 +1,9 @@
 'use strict';
 
-function print (deps, type) {
-  if (deps.length > 0) {
+const warning = (dependencies, type) => {
+  if (dependencies.length > 0) {
     console.log(`========= WARNING ${type} LICENSES ==========`);
-    deps.forEach((dep) => {
+    dependencies.forEach((dep) => {
       const licenseNames = [];
       dep.licenses.license.forEach((license) => {
         if (license.url === 'UNKNOWN') {
@@ -16,8 +16,21 @@ function print (deps, type) {
     });
     console.log(`========= WARNING ${type} LICENSES ==========`);
   }
-}
+};
+
+const print = (list, separator) => {
+  if (list.size > 0) {
+    console.log(separator);
+    Array.from(list).forEach((l) => {
+      console.log('name:', l.name,
+           ', version:', l.version,
+           ', licenses:', l.license);
+    });
+    console.log(separator);
+  }
+};
 
 module.exports = {
-  print
+  print,
+  warning
 };
