+++ lib/unified-list.js
@@ -2,28 +2,17 @@
 
 const reader = require('./file-reader.js');
 const validator = require('./json-validator.js');
+const messages = require('./messages.js');
 
-const approvedSeparator = `========= APPROVED LICENSES        ==========`;
-const notApprovedSeparator = `========= NOT APPROVED LICENSES    ==========`;
-const clarificationSeparator = `========= CLARIFICATION ADVISED    ==========`;
+const APPROVED_SEPARATOR = `========= APPROVED LICENSES        ==========`;
+const NOT_APPROVED_SEPARATOR = `========= NOT APPROVED LICENSES    ==========`;
+const CLARIFICATION_SEPARATOR = `========= CLARIFICATION ADVISED    ==========`;
 
 let unifiedList;
 
 // This function only initializes the unifiedList variable
 // to avoid do it again when call urlForName function.
 const init = (options) => { unifiedList = reader.readAsJson(options.unifiedList); };
-
-const print = (list, separator) => {
-  if (list.size > 0) {
-    console.log(separator);
-    Array.from(list).forEach((l) => {
-      console.log('name:', l.name,
-           ', version:', l.version,
-           ', licenses:', l.license);
-    });
-    console.log(separator);
-  }
-};
 
 const foundInList = (license, listItem) => {
   return (license === listItem['fedora_abbrev'] ||
@@ -106,9 +95,9 @@ function check (options, xmlObject) {
   const approvedFound = findApproved(approvedList, licensesFound);
   const notApprovedFound = findNotApproved(notApprovedList, licensesFound);
   const unknown = findUnknown(unknownList, licensesFound);
-  print(approvedFound, approvedSeparator);
-  print(notApprovedFound, notApprovedSeparator);
-  print(unknown, clarificationSeparator);
+  messages.print(approvedFound, APPROVED_SEPARATOR);
+  messages.print(notApprovedFound, NOT_APPROVED_SEPARATOR);
+  messages.print(unknown, CLARIFICATION_SEPARATOR);
 }
 
 function urlForName (options, name) {
