+++ lib/version-handler.js
@@ -1,6 +1,6 @@
 'use strict';
 
-function fromNpmVersion (version) {
+const fromNpmVersion = (version) => {
   const nameVersion = version.split('@');
   // Scoped packages might be used where the version
   // would be in the format @org/foo@1.1.0.
@@ -14,26 +14,21 @@ function fromNpmVersion (version) {
     name: nameVersion[0],
     version: nameVersion[1]
   };
-}
+};
 
-function fromNpmVersionIgnoreRange (version) {
-  return {
-    name: version.split('@')[0],
-    version: version.split('@')[1]
-  };
-}
+const fromNpmVersionIgnoreRange = (v) => ({ name: v.split('@')[0], version: v.split('@')[1] });
+const nameVersion = (name, version) => `${name}@${version}`;
 
-function asNpmVersion (name, version) {
+const asNpmVersion = (name, version) => {
   if (/^[~^]\d+/.test(version)) {
     version = version.substring(1);
   }
   return `${name}@${version}`;
-}
-
-function nameVersion (name, version) {
-  return `${name}@${version}`;
-}
+};
 
 module.exports = {
-  fromNpmVersion, asNpmVersion, nameVersion, fromNpmVersionIgnoreRange
+  fromNpmVersion,
+  asNpmVersion,
+  nameVersion,
+  fromNpmVersionIgnoreRange
 };
